sub/requirements.txt
alembic==1.16.2
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.2.1
Flask==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
nats-py==2.10.0
netifaces==0.11.0
peewee==3.18.2
prometheus_client==0.21.1
Pygments==2.19.2
PyYAML==6.0.2
requests==2.32.3
scapy==2.6.1
setuptools==80.9.0
soupsieve==2.6
SQLAlchemy==2.0.41
sqlite-web==0.6.4
typing_extensions==4.14.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
psycopg[binary]
